Objects are now fully created from JSON Some set up for backend calls and styling now too
