+++ index.html
@@ -3,6 +3,7 @@
 	<head>
 		<meta charset="utf-8">
 		<title>Course Scheduler</title>
+		<link rel="stylesheet" type="text/css" href="./src/styles/main.css">
 	</head>
 	<body>
 		<!-- This is where the entire view will be rendered -->
