+++ src/styles/main.css
@@ -1,0 +1,9 @@
+left {
+	float: left;
+	margin: 10px 10px;
+}
+
+right {
+	float: right;
+	margin: 10px 10px;
+}
