+++ src/views/overallScheduleView.tsx
@@ -2,31 +2,38 @@
 import * as React from "react";
 import {OverallSchedule, IOverallSchedule} from "../scripts/OverallSchedule"
 import SemesterView from "./semesterView"
+import {Semester, ISemester} from "../scripts/Semester"
 
-class OverallScheduleView extends React.Component<any, any> {
-
-	constructor(props: any) {
+class OverallScheduleView extends React.Component<any, any> 
+{
+	constructor(props: IOverallSchedule) 
+	{
     super(props);
 		// Convert the passed data to a parsed JSON make it an OverallSchedule
-		this.state = { aSchedule: OverallSchedule.fromJson(JSON.parse(this.props.aSchedule))};
+		this.state = {aSchedule: this.props.aSchedule};
+		console.log("Overall Schedule test");
+		console.log(this.state.aSchedule);
   }
 
   render() 
 	{
-		let semesterList = this.state.aSchedule.theSemesters.map(function(aSemester)
-			{
-				return <SemesterView aSemester={aSemester}/>;
-			})
+		let semesterList: Semester[] = this.state.aSchedule.theSemesters.map(function(aSemester)
+		{
+			// console.log("Listing Semesters");
+			// console.log(aSemester);
+			return <SemesterView aSemester={aSemester}/>;
+		});
 
 		return(
-		<div>
-			<p>Major: {this.state.aSchedule.theMajors[0].theName}</p>
-			<p>Minor: {this.state.aSchedule.theMinors[0].theName}</p>
-			<p>Semesters:</p>
-			<ul>
-				{semesterList}
-			</ul>
-		</div>
+			<div>
+				<p>Major: {this.state.aSchedule.theMajors[0].theName}</p>
+				<p>Minor: {this.state.aSchedule.theMinors[0].theName}</p>
+				<p> Total Credits: {this.state.aSchedule.theCredits}</p>
+				<p>Semesters:</p>
+				<ul>
+					{semesterList}
+				</ul>
+			</div>
 		);
 
   }
